drm/i915: set dev_priv->fbc.crtc before scheduling the enable work
This thing where we need to get the crtc either from the work
structure or the fbc structure itself is confusing and unnecessary.
Set fbc.crtc right when scheduling the enable work so we can always
use it.

Besides the immediate advantages, this is also going to make one of
the next commits much simpler.

Signed-off-by: Paulo Zanoni <paulo.r.zanoni@intel.com>