drm/i915: introduce is_active/activate/deactivate to the FBC terminology
The long term goal is to have enable/disable as the higher level
functions and activate/deactivate as the lower level functions, just
like we do for PSR and for the CRTC. Let's start this by renaming the
functions that touch the hardware state and their wrappers.

The conversion to the full correct terminology is not complete yet, so
there is still some confusion that will be solved in the next patches.

Signed-off-by: Paulo Zanoni <paulo.r.zanoni@intel.com>