drm/i915: introduce intel_fbc_{enable, disable}
These are the functions that lock/unlock FBC to/from a CRTC without
changing the hardware state. The goal is to run once per modeset. For
now the functions don't really do much, but we'll still move more
stuff there, such as the stolen memory allocations.

With this, activate/deactivate will just start/stop FBC without
changing its associated pipe. This is the same naming scheme we use in
other features, such as PSR.

This will also help in case we decide to move FBC to pipe_config or
something else later.

Signed-off-by: Paulo Zanoni <paulo.r.zanoni@intel.com>