staging/wilc1000: turn enable_irq/disable_irq into callbacks
As a preparation for turning the SDIO side of wilc1000 into a separate
module, this removes the last direct caller from the core module into
the sdio specific portion. All calls to wilc1000_sdio_enable_interrupt()
and wilc1000_sdio_disable_interrupt() now go through a function pointer
in wilc1000_ops. We also change arguments slightly to pass the device,
as we are already touching those lines and the change will be needed
later to remove the global variables.

The linux_wlan_sdio.h file is now unused and gets deleted.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>