KVM: x86: fix eflags state following processor init/reset
Reference SDM 3.4.3:

Following initialization of the processor (either by asserting the
RESET pin or the INIT pin), the state of the EFLAGS register is
00000002H.

However, the eflags fixed bit is not set and other bits are also not
cleared during the init/reset in kvm.

This patch fix it by set eflags register to 00000002H following
initialization of the processor.

Signed-off-by: Wanpeng Li <wanpeng.li@hotmail.com>