sched/deadline: fix earliest_dl.next is not the pushable candidate
earliest_dl.next is used to cache the next earliest ready task
which also is pushable in order to be a candidate of pushable
tasks during pull algorithm. If the earliest_dl.next deadline
of the sr_rq is earlier than the earliest_dl.curr deadline of
current rq, the task from the sr_rq can be pulled. However,
current implementation logic of earliest_dl.next just guarantee
it is the next earliest ready task instead of the next pushable
earliest task, which will result in hold both rqs' lock and find
nothing to preempt our current since the affinity in pull algorithm.
In addition, current logic doesn't update the next candidate for
pushing in pick_next_task_dl() even if the running task is never
eligible.

This patch fix it by updating earliest_dl.next when pushable dl task is
enqueued/dequeued the same way as rt class.

Signed-off-by: Wanpeng Li <wanpeng.li@hotmail.com>