cpufreq: imx6q: check regulator_get_optional return value
While pu_reg is an optional regulator we still have to check
the return value for -EPROBE_DEFER to properly detect the
case when we have a pu regulator but it is not yet present.

While at it, set pu_reg to NULL for an erroneous regulator
so that we can check for the regulator with if (pu_reg)
instead of the slightly less readable IS_ERR().

Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>