e1000: free skb when returning on -ENOMEM error
eth_skb_pad returns 0 if it was successful, or -ENOMEM if it was not. In
that case, this function exits early. Some early exits return with
NETDEV_TX_BUSY, which queues the skb up to be tried again, and so the
skb should not be freed. Other early exits return with NETDEV_TX_OK,
like this one, in which case it's imperative that the skb is freed,
since it is not queued back up. In this case, upon receiving -ENOMEM
from eth_skb_pad, the function exits early with NETDEV_TX_OK, but
forgets to free the skb. This patch fixes that.

In a low memory situation, in which the GFP_ATOMIC allocation from
eth_skb_pad fails, if a network device is transmitting repeatedly, this
bug could lead to rapidly leaking memory that could only be recovered by
a reboot or crash.

Signed-off-by: Jason A. Donenfeld <Jason@zx2c4.com>