power: qcom_smbb: Add otg regulator for control of vbus
Add a regulator to control the OTG chargepath switch.  The OTG
switch gets its power from pm8941_5vs1, and that should be expressed
as an usb_otg_in-supply property in the DT node for the charger driver.
The regulator name is "otg".  This is used by USB code to control VBUS
direction.

Signed-off-by: Tim Bird <tim.bird@sonymobile.com>