cfg80211: Enable update of properties for unassociated station.
Currently, cfg80211 rejects updates of AID and listen interval parameters
for existing entries. This information is known only at association stage,
and as a result it's impossible to update entries that were added
unassociated.
Fix this by allowing updates of these properies for stations that the
driver (or mac80211) assigned unassociated state.

Signed-off-by: Ayala Beker <ayala.beker@intel.com>
Signed-off-by: Emmanuel Grumbach <emmanuel.grumbach@intel.com>