blk-mq: avoid excessive boot delays with large lun counts
Hi,

Zhangqing Luo reported long boot times on a system with thousands of
LUNs when scsi-mq was enabled.  He narrowed the problem down to
blk_mq_add_queue_tag_set, where every queue is frozen in order to set
the BLK_MQ_F_TAG_SHARED flag.  Each added device will freeze all queues
added before it in sequence, which involves waiting for an RCU grace
period for each one.  We don't need to do this.  After the second queue
is added, only new queues need to be initialized with the shared tag.
We can do that by percolating the flag up to the blk_mq_tag_set, since
new hctxs already inherit the tag_set's flags.  That's what the below
patch does.

The re-check in blk_mq_add_queue_tag_set is done because we could have
two queues racing to be the second queue added for the tag set.  In such
a case, it's possible that not all of the hctxs within the loser will be
initialized.  Because this initialization is done sequentially, we can
simply check for the BLK_MQ_F_TAG_SHARED flag in the first hctx.

This problem was introduced by commit 0d2602ca30e41 (blk-mq: improve
support for shared tags maps).

Reported-and-tested-by: Jason Luo <zhangqing.luo@oracle.com>
Signed-off-by: Jeff Moyer <jmoyer@redhat.com>